Add constructor parameter to Scanner to control whether atexit is registered.
Add logout method to Scanner.

+++ nessrest/ness6rest.py
@@ -57,7 +57,7 @@ class Scanner(object):
     Scanner object
     '''
     def __init__(self, url, login='', password='', api_akey='', api_skey='',
-                 insecure=False, ca_bundle=''):
+                 insecure=False, ca_bundle='', auto_logout=True):
         self.api_akey = None
         self.api_skey = None
         self.use_api = False
@@ -120,8 +120,9 @@ class Scanner(object):
             self._login(login, password)
 
             # Register a call to the logout action automatically
-            atexit.register(self.action, action="session",
-                            method="DELETE", retry=False)
+            if auto_logout:
+                atexit.register(self.action, action="session",
+                                method="DELETE", retry=False)
 
         self._get_permissions()
         self._get_scanner_id()
@@ -144,6 +145,10 @@ class Scanner(object):
             if self.res["error"]:
                 raise Ness6RestException("It looks like you're trying to login into a Nessus 5 instance.")
 
+
+################################################################################
+    def logout(self):
+        self.action(action="session", method="delete", retry=False)
 
 ################################################################################
     def _get_permissions(self):
